shopTests created

+++ Code/src/main/java/Behaviors/ISell.java
@@ -2,5 +2,5 @@ package behaviors;
 
 public interface ISell {
 
-    int calculateMarkup();
+    double calculateMarkup();
 }
