+++ Code/src/main/java/instruments/Guitar.java
@@ -4,7 +4,7 @@ public class Guitar extends Instrument{
 
     private int numStrings;
 
-    public Guitar(String type, String colour, String material, int buyPrice, int sellPrice, int numStrings ) {
+    public Guitar(String type, String colour, String material, double buyPrice, double sellPrice, int numStrings ) {
         super(type, colour, material, buyPrice, sellPrice);
         this.numStrings = numStrings;
     }
