+++ Code/src/main/java/instruments/Piano.java
@@ -4,7 +4,7 @@ public class Piano extends Instrument{
 
     private int numKeys;
 
-    public Piano(String type, String colour, String material, int buyPrice, int sellPrice, int numKeys ) {
+    public Piano(String type, String colour, String material, double buyPrice, double sellPrice, int numKeys ) {
         super(type, colour, material, buyPrice, sellPrice);
         this.numKeys = numKeys;
     }
