+++ Code/src/main/java/instruments/Trumpet.java
@@ -2,7 +2,7 @@ package instruments;
 
 public class Trumpet extends Instrument{
 
-    public Trumpet(String type, String colour, String material, int buyPrice, int sellPrice) {
+    public Trumpet(String type, String colour, String material, double buyPrice, double sellPrice) {
         super(type, colour, material, buyPrice, sellPrice);
     }
